Translations does not stick when navigating Fixes #107
