+++ src/app/page.tsx
@@ -1,126 +1,21 @@
 "use client";
-import { Fragment, useState } from "react";
-import Wizrad from "@/app/[lng]/Components/Wizard/Wizard";
-import WelcomeMsg from "@/app/[lng]/Components/UI/WelcomeMsg";
-import { Provider } from "react-redux";
-import store from "@/store";
-import { Filter } from "./[lng]/general/interfaces";
-import Gallary from "./[lng]/Components/gallary/gallary";
-import { GetFilters } from "@/app/[lng]/general/utils";
-import { LocaleTypes } from "@/i18n/settings";
-import TrialAccordion from "@/app/[lng]/Components/ResultComp/CustomAccordion";
-import { Content } from "@/app/[lng]/general/interfaces";
+import { Typography, Link, Box } from "@mui/material";
+import { useRouter } from "next/navigation";
 
-const dummyData: Content[] = [
-    {
-        title: "Dummy Title 1",
-        link: "https://dummylink1.com",
-        tags: [
-            {
-                id: "1",
-                display: "tags1",
-                used: true,
-                languageId: "en",
-            },
-            {
-                id: "2",
-                display: "tags2",
-                used: true,
-                languageId: "en",
-            },
-        ],
-        organization: {
-            id: "1",
-            display: "Dummy Organization 1",
-            used: true,
-            languageId: "en",
-        },
-        description: "This is a dummy description 1",
-        languageId: "en",
-        uploader: "Dummy Uploader 1",
-        id: "1",
-    },
-    {
-        title: "Dummy Title 2",
-        link: "https://dummylink2.com",
-        tags: [
-            {
-                id: "3",
-                display: "tags3",
-                used: true,
-                languageId: "en",
-            },
-            {
-                id: "4",
-                display: "tags4",
-                used: true,
-                languageId: "en",
-            },
-        ],
-        organization: {
-            id: "2",
-            display: "Dummy Organization 2",
-            used: true,
-            languageId: "en",
-        },
-        description: "This is a dummy description 2",
-        languageId: "fr",
-        uploader: "Dummy Uploader 2",
-        id: "2",
-    },
-    {
-        title: "Dummy Title 3",
-        link: "https://dummylink3.com",
-        tags: [
-            {
-                id: "5",
-                display: "tags5",
-                used: true,
-                languageId: "en",
-            },
-            {
-                id: "6",
-                display: "tags6",
-                used: true,
-                languageId: "en",
-            },
-        ],
-        organization: {
-            id: "3",
-            display: "Dummy Organization 3",
-            used: true,
-            languageId: "en",
-        },
-        description:
-            "This is a dummy description 3 This is a dummy description 3 This is a dummy description 3 This is a dummy description 3 This is a dummy description 3 This is a dummy description 3 ",
-        languageId: "es",
-        uploader: "Dummy Uploader 3",
-        id: "3",
-    },
-];
-
-function HomePage({ params: { lng } }: { params: { lng: LocaleTypes } }) {
-    const [open, setOpen] = useState(false);
-
-    const openWizradHandler = () => {
-        setOpen(true);
-    };
-
-    const closeWizradHandler = () => {
-        setOpen(false);
-    };
-
-    const filters: Filter[] = GetFilters();
-
+export default function RedirectingPage() {
+    const router = useRouter();
+    router.replace("he/home-page");
     return (
-        <Provider store={store}>
-            <Fragment>
-                <Wizrad open={open} onClose={closeWizradHandler} />
-                <WelcomeMsg lng={lng} openWizradHandler={openWizradHandler} />
-                <Gallary filters={filters} />
-            </Fragment>
-        </Provider>
+        <Box textAlign="center" sx={{ mt: "25%" }}>
+            <Typography variant="h2" color="black">
+                ...Redirecting
+            </Typography>
+            <br />
+            <Typography variant="h4" color="black" align="center">
+                If you are not redirected automatically, follow this
+                <br />
+                <Link href="/he/home-page">link</Link>
+            </Typography>
+        </Box>
     );
 }
-
-export default HomePage;
